update settings view to Split Location & Calculation into separate Location and Calculation screens
