+++ Salah/Resources/Localizable.xcstrings
@@ -2526,6 +2526,7 @@
       }
     },
     "Location & Calculation" : {
+      "extractionState" : "stale",
       "localizations" : {
         "bn" : {
           "stringUnit" : {
